[WCNCR00276737][Coverity code defect]
Invalid type in argument to printf format specifier

MTK-Commit-Id: 14fb7b10f0255b8a413faa05b98f722811861c54

CR-Id: WCNCR00276737
Reviewed-on: https://gerrit.mediatek.inc/c/neptune/wlan_driver/gen4m/+/6891546
CheckForbiddenKeyword: srv_pis <srv_pis@mediatek.com>
Build: srv_preflight_a001 <srv_preflight_a001@mediatek.com>
CheckPatch: srv_pis <srv_pis@mediatek.com>
Build: srv_neptune_adm <srv_neptune_adm@mediatek.com>
Test: srv_pf_nep_sanity
CheckCRInfo: srv_pis <srv_pis@mediatek.com>
CheckSWIP: srv_pis <srv_pis@mediatek.com>
CheckCourse: srv_pis <srv_pis@mediatek.com>

+++ mgmt/twt.c
@@ -457,8 +457,8 @@ static void btwtParseTWTElement(
 
 	DBGLOG(TWT_REQUESTER, WARN,
 		"BTWT cur TSF %x %x\nAP resp TWT %x\n",
-		(prTWTParams->u8TWT & 0x00000000FFFFFFFF),
-		((prTWTParams->u8TWT & 0xFFFFFFFF00000000) >> 32),
+		(uint32_t)(prTWTParams->u8TWT & 0x00000000FFFFFFFF),
+		(uint32_t)((prTWTParams->u8TWT & 0xFFFFFFFF00000000) >> 32),
 		prBTWTIE->u2TWT);
 
 	u8TargetWakeTime = prTWTParams->u8TWT;
@@ -470,8 +470,8 @@ static void btwtParseTWTElement(
 
 	DBGLOG(TWT_REQUESTER, WARN,
 		"BTWT twt %x %x\n",
-		(prTWTParams->u8TWT & 0x00000000FFFFFFFF),
-		((prTWTParams->u8TWT & 0xFFFFFFFF00000000) >> 32));
+		(uint32_t)(prTWTParams->u8TWT & 0x00000000FFFFFFFF),
+		(uint32_t)((prTWTParams->u8TWT & 0xFFFFFFFF00000000) >> 32));
 }
 
 uint8_t btwtGetTxSetupFlowId(
